refactor: errorBox gets children props instead error, export/import
