+++ src/components/ErrorBox/ErrorBox.jsx
@@ -1,10 +1,10 @@
 import PropTypes from "prop-types";
 import styles from "./errorBox.module.scss";
 
-export const ErrorBox = ({ error }) => {
-  return <span className={styles.error}>{error}</span>;
+export const ErrorBox = ({ children }) => {
+  return <span className={styles.error}>{children}</span>;
 };
 
 ErrorBox.propTypes = {
-  error: PropTypes.string.isRequired,
+  children: PropTypes.node.isRequired,
 };
