+++ src/components/index.js
@@ -8,3 +8,4 @@ export { Logo } from "./Logo/Logo";
 export { Navbar } from "./Navbar/Navbar";
 export { SearchForm } from "./SearchForm/SearchForm";
 export { Select } from "./Select/Select";
+export { ErrorBox } from "./ErrorBox/ErrorBox";
